backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
astroid==3.3.8
bcrypt==4.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dill==0.3.9
fastapi==0.115.8
h11==0.14.0
idna==3.10
isort==6.0.0
mccabe==0.7.0
neo4j==5.28.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
ruff==0.9.6
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
